Fix: Only invoking HighlightingStateChanged for changes lines.
Found that DotNetPad was getting slower to type in when number of lines
increased. Since HighlightLine in CodeHighlighter is called for the
changed line plus all lines above it there is a lot of rendering going
on.

This fix changes so that HighlightingStateChanged is only invoked if all
sections on a line is equal to the old version of the same line.